improved ProcessBar Performance
Added a minimum delay betweeen two statusupdates.

This creates difference in speed which gets bigger
while the number of fast triggered updates gets bigger.
This happens for example if you trigger updates after
every processed line of an input-file.

The Delay is set by the new __construct-parameter $framesPerSecound
and is used with a defaultvalue of 15 frames (one every 0.06666 seconds).

Signed-off-by: David Sauer <david@nur-mal-so-am-ran.de>